Fix compiler warning caused by not using "unsigned" keyword in DIS 7

+++ src/dis7/EntityDamageStatusPdu.cpp
@@ -135,7 +135,7 @@ int EntityDamageStatusPdu::getMarshalledSize() const
    marshalSize = marshalSize + 2;  // _padding2
    marshalSize = marshalSize + 2;  // _numberOfDamageDescription
 
-   for(int idx=0; idx < _damageDescriptionRecords.size(); idx++)
+   for(unsigned long idx=0; idx < _damageDescriptionRecords.size(); idx++)
    {
         DirectedEnergyDamage listElement = _damageDescriptionRecords[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
